Working mining

+++ src/main/kotlin/com/innoq/chainy/model/Status.kt
@@ -2,4 +2,4 @@ package com.innoq.chainy.model
 
 import java.util.*
 
-data class Status(val nodeId: UUID, val currentBlockHeight: Long)
+data class Status(val nodeId: UUID, val currentBlockHeight: Int)
